fix: Don't try to use colors on non-terminals
If the TERM environment variable is never set, we probably aren't in a
terminal environment at all, and we shouldn't try to use colors.

Still fall back to indexed 16 if a terminal is present and we simply
don't recognize it, as all terminals should at least support the bare
minimum.